+++ src/main/java/net/silentchaos512/mechanisms/config/OreConfig.java
@@ -1,32 +1,29 @@
 package net.silentchaos512.mechanisms.config;
 
+import net.minecraftforge.common.ForgeConfigSpec;
 import net.silentchaos512.mechanisms.init.Ores;
-import net.silentchaos512.utils.config.BooleanValue;
-import net.silentchaos512.utils.config.ConfigSpecWrapper;
-import net.silentchaos512.utils.config.IntValue;
 
 public class OreConfig {
-    private final BooleanValue masterSwitch;
-    private final IntValue veinCount;
-    private final IntValue veinSize;
-    private final IntValue minHeight;
-    private final IntValue maxHeight;
+    private final ForgeConfigSpec.BooleanValue masterSwitch;
+    private final ForgeConfigSpec.IntValue veinCount;
+    private final ForgeConfigSpec.IntValue veinSize;
+    private final ForgeConfigSpec.IntValue minHeight;
+    private final ForgeConfigSpec.IntValue maxHeight;
 
-    public OreConfig(Ores ore, ConfigSpecWrapper wrapper, BooleanValue masterSwitch) {
-        String key = "world.gen." + ore + ".";
+    public OreConfig(Ores ore, ForgeConfigSpec.Builder builder, ForgeConfigSpec.BooleanValue masterSwitch) {
         this.masterSwitch = masterSwitch;
-        this.veinCount = wrapper.builder(key + "veinCount")
+        this.veinCount = builder
                 .comment("Number of veins per chunk")
-                .defineInRange(ore.getDefaultOreConfigs().getVeinCount(), 0, Integer.MAX_VALUE);
-        this.veinSize = wrapper.builder(key + "veinSize")
+                .defineInRange(ore.getName() + ".veinCount", ore.getDefaultOreConfigs().getVeinCount(), 0, Integer.MAX_VALUE);
+        this.veinSize = builder
                 .comment("Size of veins")
-                .defineInRange(ore.getDefaultOreConfigs().getVeinSize(), 0, 100);
-        this.minHeight = wrapper.builder(key + "minHeight")
+                .defineInRange(ore.getName() + ".veinSize", ore.getDefaultOreConfigs().getVeinSize(), 0, 100);
+        this.minHeight = builder
                 .comment("Minimum Y-coordinate (base height) of veins")
-                .defineInRange(ore.getDefaultOreConfigs().getMinHeight(), 0, 255);
-        this.maxHeight = wrapper.builder(key + "maxHeight")
+                .defineInRange(ore.getName() + ".minHeight", ore.getDefaultOreConfigs().getMinHeight(), 0, 255);
+        this.maxHeight = builder
                 .comment("Maximum Y-coordinate (highest level) of veins")
-                .defineInRange(ore.getDefaultOreConfigs().getMaxHeight(), 0, 255);
+                .defineInRange(ore.getName() + ".maxHeight", ore.getDefaultOreConfigs().getMaxHeight(), 0, 255);
     }
 
     public boolean isEnabled() {